requirements.txt
# FastAPI and server
fastapi
uvicorn[standard]
python-multipart
transformers
torch
torchvision
Pillow
numpy
scipy